Support complex GOPATH.
Issue #68 points out that Glide needs to support the case where a
GOPATH contains a colon-separated list of paths.

This patch provides two functions, Gopath() and Gopaths(), for assisting
with this situation. Gopath() returns the most relevant Gopath()
(e.g. the first one on a multi-path set). Gopaths() returns all
of the paths.

I've adjusted some read operations to use Gopaths(), and elsewhere
used Gopath(). This is not always a straightforward decision, so
there may be bugs with complex Gopath situations.