Adding fastpath update back to Glide
Fastpath updates skip pulling updates if the the current set version
is the version already checked out. For pinned updates or those
set to a tag this will make updates faster because for DVCS cases
glide doesn't need to go out to the Internet. Note, when branches
are set at the Reference updates still occur.