Update stubbed out and partially functional.
The implementation of update here is much different than the old one.
It does only one pass over the codebase instead of two, and it
parallelizes git fetch and version setting.

Currently, it does not have all of the requisite package handling done,
and it is not correctly setting versions.