Cleaned up per go vet
Note, this change caused a public API change. msg.Error became
msg.Err. Go vet expects Error functions to not have formatting. If
they do it expects Errorf. This is even for non-standard library
functions. Changing to Err because it aligns with Info and Warn
which are both short hand names.