Fixed #272: Ignore appengine packages only available when built in appengine
Appengine has an appengine package (with sub-packages) that are only
available when build inside of appengine. They are in the same space
as standard library packages. Now Glide detects them and the resolver
handles them appropriately (skips) them. If the application is built
in appengine the package will be there just like a standard library
one.