Fixed #273: Glide fails when multiple packages in same directory
Go does not allow multiple packages in the same directory. But,
some use build tags to ignore example files. We need to look at
dependencies for all OS which is why we ignore build tags by
default. When ignoring build tags if we run into an error
of multiple packages we iterate over all possible combinations
to generate the list.