Glide install skips fetch when it is up to date
For 'glide install', this lazily computes whether the existing vendor
checkouts are already at the right commit. If they are, those deps are
skipped during the concurrent update.

This is safe to do in install because the dependency graph is already
computed, so there should never be a case where the glide.lock file does
not contain a negotiable dependency graph.