Add RPM spec file and version management to Makefile.
This commit allows user to build an RPM package with a single command just
using source tarball. To achieve this i've added an sdist target to the
Makefile. This target will generate source tarball and put VERSION file inside
to allow make utility to detect version of the source tarball even if no git is
present.

If used a tarball generated by sdist target it is possible to build RPMs with a
single command:

    rpmbuild --define "_glide_version 0.11.1-11-gb5ebdcb" -ta \
        dist/glide-0.11.1-11-gb5ebdcb-src.tar.gz