Carve out the other references to legacy config
This is a really, really awkward hack job. I'm certain that this breaks
at least some of the functionality here in ways that aren't necessary.
It was simply too much work to grok all of these changes in context;
it'll be necessary to circle back around on them as part of the later
porting work.

I've placed TODOs/FIXMEs at some particularly crucial spots to assist
with this.