Glide Cache Path
This patch introduces a new command, "cache". The "cache" command
has two sub-commands: "clear" and "path". The "cache clear" command
is simply the former "cache-clear" command relocated under the new
"cache" command. The "path" command may be used to return the location
of a dependency's sources in the Glide cache. For example:

        $ cd $GOPATH/src/github.com/emccode/rexray && \
          glide cache path github.com/emccode/libstorage

The above command would return the following location:

        $HOME/.glide/cache/src/https-github.com-emccode-libstorage