Use cache directory as defined by XDG Base Directory Specification
From the specification:

https://specifications.freedesktop.org/basedir-spec/basedir-spec-0.8.html

There is a single base directory relative to which user-specific
non-essential (cached) data should be written. This directory is defined
by the environment variable $XDG_CACHE_HOME.

$XDG_CACHE_HOME defines the base directory relative to which user specific
non-essential data files should be stored. If $XDG_CACHE_HOME is either
not set or empty, a default equal to $HOME/.cache should be used.